Caster代表Buff的施加者，它有可能为空，也有可能不为空，视具体构造时是否传Caster参数而定。但是Buff有一个配置项bNoCaster（是否强制设置Caster为空）。如果bNoCaster = true。则Buff的Caster一定为空。

+++ node/game/src/game_logic/combat/buff/comp/buff_comp.h
@@ -21,5 +21,5 @@ struct BuffComp
 
 struct BuffListComp
 {
-    std::map<uint32_t, BuffComp> buffList;
+    std::map<uint64_t, BuffComp> buffList;
 };
